Switch out the WorldGateManager for modsystem
